refactor: extract STATE to memory module
