+++ src/mission_control/src/memory.rs
@@ -1,0 +1,6 @@
+use crate::types::state::State;
+use std::cell::RefCell;
+
+thread_local! {
+    pub static STATE: RefCell<State> = RefCell::default();
+}
